refactor: add Quailifier to Hilt Module
