+++ app/src/main/java/campus/tech/kakao/map/di/DatabaseModule.kt
@@ -1,8 +1,12 @@
 package campus.tech.kakao.map.di
 
 import android.content.Context
+import androidx.datastore.core.DataStore
+import androidx.datastore.dataStore
 import androidx.room.Room
 import campus.tech.kakao.map.data.database.AppDatabase
+import campus.tech.kakao.map.data.model.Location
+import campus.tech.kakao.map.data.repository.LocationSerializer
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -13,8 +17,24 @@ import dagger.hilt.android.scopes.ViewModelScoped
 @Module
 @InstallIn(ViewModelComponent::class)
 object DatabaseModule {
+
+    private val Context.dataStore: DataStore<Location> by dataStore(
+        fileName = "location_data.pb",
+        serializer = LocationSerializer,
+    )
+
     @Provides
     @ViewModelScoped
+    @LocationDataStore
+    fun provideDataStore(
+        @ApplicationContext context: Context,
+    ): DataStore<Location> {
+        return context.dataStore
+    }
+
+    @Provides
+    @ViewModelScoped
+    @SearchWordDatabase
     fun provideDatabase(
         @ApplicationContext context: Context,
     ): AppDatabase {
@@ -23,5 +43,6 @@ object DatabaseModule {
 
     @Provides
     @ViewModelScoped
-    fun provideSavedSearchWordDao(database: AppDatabase) = database.savedSearchWordDao()
+    fun provideSavedSearchWordDao(@SearchWordDatabase database: AppDatabase) = database.savedSearchWordDao()
+
 }
