+++ app/src/main/java/campus/tech/kakao/map/di/NetworkModule.kt
@@ -5,14 +5,12 @@ import campus.tech.kakao.map.data.network.service.KakaoLocalService
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ViewModelComponent
-import dagger.hilt.android.scopes.ViewModelScoped
+import dagger.hilt.components.SingletonComponent
 
 @Module
-@InstallIn(ViewModelComponent::class)
+@InstallIn(SingletonComponent::class)
 object NetworkModule {
     @Provides
-    @ViewModelScoped
     fun provideKakaoLocalService(): KakaoLocalService {
         return KaKaoLocalApiClient.retrofit.create(KakaoLocalService::class.java)
     }
